add jobq
A jobq is a refactored version of a workq which primarily
removes the map_task since this can cause performance
issues in some cases. A jobq is intended to be used for
cases when a task will be run once and is expected to run
to completion. A workq is intended to be used for cases
when a task may need to be canceled/purged before
completion.